+++ src/App.js
@@ -1,6 +1,7 @@
 import './App.css';
 import { Component } from 'react';
-import { PostCard } from './components/PostCard';
+import { loadPosts } from './utils/load-posts'
+import { Posts } from './components/Posts';
 
 class App extends Component {
 
@@ -13,47 +14,16 @@ class App extends Component {
   }
 
   loadPosts = async () => {
-    const postsResponse = 
-      fetch('https://jsonplaceholder.typicode.com/posts');
-
-    const phtosResponse = 
-      fetch('https://jsonplaceholder.typicode.com/photos');
-
-    const [posts, photos] = 
-      await Promise.all([postsResponse, phtosResponse]);
-
-    const postsJson = await posts.json();
-    const photosJson = await photos.json();
-
-    const postsAndPhotos  = postsJson.map((post, index) => {
-      return { ...post, cover: photosJson[index].url }
-    });  
-
-    this.setState({posts: postsAndPhotos});
-    
+    const postsAndPhotos = await loadPosts();
+    this.setState({ posts: postsAndPhotos })
   }
 
-  componentDidUpdate() {
-  }
-
-  componentWillInmount() {
-  }
 
   render() {
     const { posts } = this.state;
     return (
       <section className='container'>
-        <div className="posts">
-          {posts.map(post => (
-            <PostCard
-              key={post.id}  
-              title={post.title} 
-              body={post.body}
-              id={post.id}
-              cover={post.cover}
-            />
-          ))}
-      </div>
+       <Posts posts={posts} />
      </section>
     );
   }
