+++ src/components/PostCard/index.jsx
@@ -1,9 +1,8 @@
 export const PostCard = ({title, body, cover, id}) => (
     <div className="post">
         <img src={cover} alt={title} />
-        <div
-        className="post-content"> 
-        <h1>{title}</h1>
+        <div className="post-content"> 
+        <h2>{title}</h2>
         <p>{body}</p>
         </div>
     </div>
