+++ src/components/Posts/index.jsx
@@ -1,0 +1,14 @@
+import { PostCard } from '../PostCard';
+export const Posts = ({ posts }) => (
+    <div className="posts">
+    {posts.map(post => (
+      <PostCard
+        key={post.id}  
+        title={post.title} 
+        body={post.body}
+        id={post.id}
+        cover={post.cover}
+      />
+    ))}
+</div>
+)
