+++ src/utils/load-posts.js
@@ -1,0 +1,19 @@
+export const loadPosts = async () => {
+    const postsResponse = 
+        fetch('https://jsonplaceholder.typicode.com/posts');
+
+    const phtosResponse = 
+        fetch('https://jsonplaceholder.typicode.com/photos');
+
+    const [posts, photos] = 
+        await Promise.all([postsResponse, phtosResponse]);
+
+    const postsJson = await posts.json();
+    const photosJson = await photos.json();
+
+    const postsAndPhotos  = postsJson.map((post, index) => {
+        return { ...post, cover: photosJson[index].url }
+    }); 
+
+    return postsAndPhotos;
+}
